# Conflicts:
#	src/main/resources/templates/View/User/Basic/LookUpTicket.html

+++ src/main/resources/static/js/RouteManagement.js
@@ -1,7 +1,7 @@
 (function () {
     document.addEventListener("DOMContentLoaded", function () {
         // start
-        modalController();
+        // modalController();
         deleteController();
         getDetailsController();
         searchController();
@@ -11,32 +11,35 @@
             deleteBtn.forEach(btn => {
                 btn.addEventListener("click", function () {
                     console.log(btn.dataset.id);
-                    fetch("http://localhost:8080/admin/routes/delete", {
-                        method: "DELETE",
-                        headers: {
-                            "Content-Type": "application/json"
-                        },
-                        body: JSON.stringify({
-                            routeId: btn.dataset.id
+                    const message = `Bạn có chắc muốn xóa tuyến đường `+btn.dataset.id+`) không?`;
+                    if (confirm(message)) {
+                        fetch("http://localhost:8080/admin/routes/delete", {
+                            method: "DELETE",
+                            headers: {
+                                "Content-Type": "application/json"
+                            },
+                            body: JSON.stringify({
+                                routeId: btn.dataset.id
+                            })
                         })
-                    })
-                        .then(response => {
-                            if (!response.ok) {
-                                throw new Error('Delete failed');
-                            }
-                            return response.json();
-                        })
-                        .then(data => {
-                            if (data === true) {
-                                alert("Delete successfully");
-                                btn.closest("li").remove();
-                            } else {
-                                alert("Some error while deleting");
-                            }
-                        })
-                        .catch(error => {
-                            alert("Delete failed: " + error);
-                        });
+                            .then(response => {
+                                if (!response.ok) {
+                                    throw new Error('Delete failed');
+                                }
+                                return response.json();
+                            })
+                            .then(data => {
+                                if (data === true) {
+                                    alert("Delete successfully");
+                                    window.location.href = "/admin/routes"
+                                } else {
+                                    alert("Some error while deleting");
+                                }
+                            })
+                            .catch(error => {
+                                alert("Delete failed: " + error);
+                            });
+                    }
                 });
             });
         }
@@ -96,54 +99,54 @@
             });
         }
 
-        function modalController() {
-            const addUserOpenBtn = document.querySelector(".add-user-open");
-            const addUserCloseBtn = document.querySelector(".add-user-close");
-            const modalContainer = document.querySelector(".modal-container");
-            const modal = modalContainer.querySelector(".modal");
-            const addUserForm = document.querySelector("#add-user-form");
-            const formSubmitBtn = modal.querySelector(".add-user-submit");
-            let isMouseHoveringForm = false;
-
-            checkMouseEnterFormZone();
-            toggleModal();
-            submitForm();
-
-            function toggleModal() {
-                addUserOpenBtn.addEventListener("click", function () {
-                    modalContainer.classList.remove("hidden");
-                    modal.classList.remove("-bottom-full");
-                });
-                addUserCloseBtn.addEventListener("click", function () {
-                    modalContainer.classList.add("hidden");
-                    modal.classList.add("-bottom-full");
-                    addUserForm.querySelectorAll("input[type='text']").forEach(input => { input.value = "" });
-                });
-                modalContainer.addEventListener("click", function () {
-                    if (isMouseHoveringForm === false) {
-                        modalContainer.classList.add("hidden");
-                        modal.classList.add("-bottom-full");
-                    }
-                });
-            }
-
-            function checkMouseEnterFormZone() {
-                modal.addEventListener("mouseenter", function () {
-                    isMouseHoveringForm = true;
-                });
-                modal.addEventListener("mouseleave", function () {
-                    isMouseHoveringForm = false;
-                });
-            }
-            function submitForm() {
-                formSubmitBtn.addEventListener("click", function () {
-                    addUserForm.submit();
-                    modalContainer.classList.add("hidden");
-                    modal.classList.add("-bottom-full");
-                    addUserForm.querySelectorAll("input[type='text']").forEach(input => { input.value = "" });
-                });
-            }
-        }
+        // function modalController() {
+        //     const addUserOpenBtn = document.querySelector(".add-user-open");
+        //     const addUserCloseBtn = document.querySelector(".add-user-close");
+        //     const modalContainer = document.querySelector(".modal-container");
+        //     const modal = modalContainer.querySelector(".modal");
+        //     const addUserForm = document.querySelector("#add-user-form");
+        //     // const formSubmitBtn = modal.querySelector(".add-user-submit");
+        //     let isMouseHoveringForm = false;
+        //
+        //     checkMouseEnterFormZone();
+        //     toggleModal();
+        //     submitForm();
+        //
+        //     function toggleModal() {
+        //         addUserOpenBtn.addEventListener("click", function () {
+        //             modalContainer.classList.remove("hidden");
+        //             modal.classList.remove("-bottom-full");
+        //         });
+        //         addUserCloseBtn.addEventListener("click", function () {
+        //             modalContainer.classList.add("hidden");
+        //             modal.classList.add("-bottom-full");
+        //             addUserForm.querySelectorAll("input[type='text']").forEach(input => { input.value = "" });
+        //         });
+        //         modalContainer.addEventListener("click", function () {
+        //             if (isMouseHoveringForm === false) {
+        //                 modalContainer.classList.add("hidden");
+        //                 modal.classList.add("-bottom-full");
+        //             }
+        //         });
+        //     }
+        //
+        //     function checkMouseEnterFormZone() {
+        //         modal.addEventListener("mouseenter", function () {
+        //             isMouseHoveringForm = true;
+        //         });
+        //         modal.addEventListener("mouseleave", function () {
+        //             isMouseHoveringForm = false;
+        //         });
+        //     }
+        //     function submitForm() {
+        //         formSubmitBtn.addEventListener("click", function () {
+        //             addUserForm.submit();
+        //             modalContainer.classList.add("hidden");
+        //             modal.classList.add("-bottom-full");
+        //             addUserForm.querySelectorAll("input[type='text']").forEach(input => { input.value = "" });
+        //         });
+        //     }
+        // }
         // end
     });
 })();
